Move console logging to logger function to avoid circular dependency in utils module
