+++ src/rocprof_compute_analyze/analysis_cli.py
@@ -25,7 +25,7 @@
 from rocprof_compute_analyze.analysis_base import OmniAnalyze_Base
 from utils import file_io, parser, tty
 from utils.kernel_name_shortener import kernel_name_shortener
-from utils.utils import console_error, demarcate
+from utils.logger import console_error, demarcate
 
 
 class cli_analysis(OmniAnalyze_Base):
