+++ src/rocprof_compute_soc/soc_base.py
@@ -37,12 +37,15 @@ import yaml
 import config
 from utils.mi_gpu_spec import get_gpu_model, get_gpu_series
 from utils.parser import build_in_vars, supported_denom
-from utils.utils import (
-    capture_subprocess_output,
+from utils.logger import (
     console_debug,
     console_error,
     console_log,
     console_warning,
+    demarcate,
+)
+from utils.utils import (
+    capture_subprocess_output,
     convert_metric_id_to_panel_idx,
     demarcate,
     detect_rocprof,
