+++ src/rocprof_compute_soc/soc_gfx906.py
@@ -26,7 +26,7 @@ from pathlib import Path
 
 import config
 from rocprof_compute_soc.soc_base import OmniSoC_Base
-from utils.utils import console_error, demarcate
+from utils.logger import console_error, demarcate
 
 
 class gfx906_soc(OmniSoC_Base):
