+++ src/utils/db_connector.py
@@ -32,13 +32,13 @@ from pymongo import MongoClient
 from tqdm import tqdm
 
 from utils.kernel_name_shortener import kernel_name_shortener
-from utils.utils import (
+from utils.utils import is_workload_empty
+from utils.logger import (
     console_debug,
     console_error,
     console_log,
     console_warning,
     demarcate,
-    is_workload_empty,
 )
 
 MAX_SERVER_SEL_DELAY = 5000  # 5 sec connection timeout
