+++ src/utils/mi_gpu_spec.py
@@ -1,8 +1,8 @@
-import logging
 import os
 import sys
 from dataclasses import dataclass, field
 from typing import Any, Dict, List, Optional, Union
+from utils.logger import console_log, console_warning, console_debug, console_error
 
 import yaml
 
@@ -96,7 +96,7 @@ class MIGPU(Singleton):
         if self.is_mi300:
             # NOTE: currently, all mi300 series gpus shall have compute partition information
             if self.compute_partition is None:
-                logging.warning(
+                console_warning(
                     "[MIGPU post init] mi300 gpu detected, but no num_xcd/compute partition data detected!!!"
                 )
 
@@ -108,7 +108,7 @@ class MIGPU(Singleton):
         All mi300 series gpus shall have compute partition information.
         """
         if num_xcds is None:
-            logging.warning(
+            console_warning(
                 "[MIGPU post init] mi300 gpu detected, but no num_xcd/compute partition data detected!!!"
             )
 
@@ -150,17 +150,17 @@ def load_yaml(file_path: str) -> Dict[str, Any]:
         Dict[str, Any]: Parsed YAML data as a nested dictionary.
                         Exit with console error if an error occurs.
     """
-    logging.debug("[load_yaml]")
+    console_debug("[load_yaml]")
     try:
         with open(file_path, "r") as file:
             data = yaml.safe_load(file)
             return data
     except FileNotFoundError:
-        logging.error(f"Error: The file '{file_path}' was not found.")
+        console_error(f"Error: The file '{file_path}' was not found.")
     except yaml.YAMLError as exc:
-        logging.error(f"Error parsing YAML file '{file_path}': {exc}")
+        console_error(f"Error parsing YAML file '{file_path}': {exc}")
     except Exception as e:
-        logging.error(
+        console_error(
             f"An unexpected error occurred while loading YAML file '{file_path}': {e}"
         )
 
@@ -186,7 +186,7 @@ def parse_mi_gpu_spec():
 
     for mi_index, mi_series in MI_CONSTANS.items():
         if mi_series != MI_CONSTANS[MI300]:
-            logging.debug("[parse_mi_gpu_spec] Processing series: %s" % mi_series)
+            console_debug("[parse_mi_gpu_spec] Processing series: %s" % mi_series)
             for key, value in yaml_data.items():
                 # parse out gpu series and gpu model information for mi50, 100, 200
                 curr_gpu_arch = value[mi_index]["gpu_archs"][0]["gpu_arch"]
@@ -235,7 +235,7 @@ def parse_mi_gpu_spec():
 
 def get_gpu_series_dict():
     if not gpu_series_dict:
-        logging.error(
+        console_error(
             "gpu_series_dict not yet populated, did you run parse_mi_gpu_spec()?"
         )
         return None
@@ -244,7 +244,7 @@ def get_gpu_series_dict():
 
 def get_gpu_series(gpu_arch_):
     if not gpu_series_dict:
-        logging.error(
+        console_error(
             "gpu_series_dict not yet populated, did you run parse_mi_gpu_spec()?"
         )
         return None
@@ -254,14 +254,14 @@ def get_gpu_series(gpu_arch_):
     if gpu_series:
         return gpu_series
 
-    logging.warning(f"No matching gpu series found for gpu arch: {gpu_arch_}")
+    console_warning(f"No matching gpu series found for gpu arch: {gpu_arch_}")
     return None
 
 
 def get_gpu_model(gpu_arch_, chip_id_):
     # Check that gpu_model_dict is populated first
     if not gpu_model_dict:
-        logging.error(
+        console_error(
             "gpu_model_dict not yet populated. Did you run parse_mi_gpu_spec()?"
         )
         return None
@@ -273,7 +273,7 @@ def get_gpu_model(gpu_arch_, chip_id_):
         if chip_id_ and int(chip_id_) in mi300_chip_id_dict:
             gpu_model = mi300_chip_id_dict.get(int(chip_id_))
         else:
-            logging.warning(f"No gpu model found for chip id: {chip_id_}")
+            console_warning(f"No gpu model found for chip id: {chip_id_}")
             return None
 
     # Otherwise use gpu_model_dict mapping for other mi architectures
@@ -281,11 +281,11 @@ def get_gpu_model(gpu_arch_, chip_id_):
         # NOTE: take the first element works for now
         gpu_model = gpu_model_dict[gpu_arch_lower][0]
     else:
-        logging.warning(f"No gpu model found for gpu arch: {gpu_arch_lower}")
+        console_warning(f"No gpu model found for gpu arch: {gpu_arch_lower}")
         return None
 
     if not gpu_model:
-        logging.warning(f"No gpu model found for gpu arch: {gpu_arch_lower}")
+        console_warning(f"No gpu model found for gpu arch: {gpu_arch_lower}")
         return None
 
     return gpu_model
@@ -293,7 +293,7 @@ def get_gpu_model(gpu_arch_, chip_id_):
 
 def get_mi300_archs_dict():
     if not mi300_archs_dict:
-        logging.error(
+        console_error(
             "mi300_archs_dict not yet populated, did you run parse_mi_gpu_spec()?"
         )
         return None
@@ -302,7 +302,7 @@ def get_mi300_archs_dict():
 
 def get_mi300_num_xcds(gpu_model_, compute_partition_):
     if not mi300_num_xcds_dict:
-        logging.error(
+        console_error(
             "mi300_num_xcds_dict not yet populated, did you run parse_mi_gpu_spec()?"
         )
         return None
@@ -311,17 +311,17 @@ def get_mi300_num_xcds(gpu_model_, compute_partition_):
     partition_lower = compute_partition_.lower()
 
     if gpu_model_lower not in mi300_num_xcds_dict:
-        logging.info(f"Current system is not a mi300 system: {gpu_model_}")
+        console_log(f"Current system is not a mi300 system: {gpu_model_}")
         return None
 
     model_dict = mi300_num_xcds_dict[gpu_model_lower]
     if partition_lower not in model_dict:
-        logging.info(f"Unknown compute partition: {compute_partition_}")
+        console_log(f"Unknown compute partition: {compute_partition_}")
         return None
 
     num_xcds = model_dict[partition_lower]
     if not num_xcds:
-        logging.warning(
+        console_warning(
             "Unknown compute partition found for %s / %s", compute_partition_, gpu_model_
         )
         return None
@@ -333,6 +333,6 @@ def get_mi300_chip_id_dict():
     if mi300_chip_id_dict:
         return mi300_chip_id_dict
     else:
-        logging.error(
+        console_error(
             "mi300_chip_id_dict not yet populated, did you run parse_mi_gpu_spec()?"
         )
