+++ src/utils/tty.py
@@ -29,7 +29,8 @@ import pandas as pd
 from tabulate import tabulate
 
 from utils import parser
-from utils.utils import console_log, console_warning, convert_metric_id_to_panel_idx
+from utils.utils import convert_metric_id_to_panel_idx
+from utils.logger import console_log, console_warning
 
 hidden_columns = ["Tips", "coll_level"]
 hidden_sections = [1900, 2000]
